'Fixed' issues with even width and height, added some error validation
Having an even number grid width or height on the DFS created a row/col of walls on the final row/col. This is due to the 2-step DFS I used but I cannot work a fix, so simply made it so that it is not possible to use even numbers.